4_spring-boot-data-jpa/src/main/java/com/example/springdatajpa/demo/repository/MemberJpaRepository.java
package com.example.springdatajpa.demo.repository;

import com.example.springdatajpa.demo.entity.Member;
import jakarta.persistence.EntityManager;
import jakarta.persistence.PersistenceContext;
import org.springframework.stereotype.Repository;

@Repository
public class MemberJpaRepository {
    @PersistenceContext
    private EntityManager em;

    public Member save(Member member) {
        em.persist(member);
        return member;
    }

    public Member find(Long id) {
        return em.find(Member.class, id);
    }
}
